+++ src/Schema/Mixin/RevisionMixinSchema.php
@@ -3,6 +3,7 @@ namespace LazyRecord\Schema\Mixin;
 use LazyRecord\Schema\MixinDeclareSchema;
 use LazyRecord\Schema;
 use DateTime;
+use SQLBuilder\Raw;
 
 class RevisionMixinSchema extends MixinDeclareSchema
 {
@@ -18,19 +19,18 @@ class RevisionMixinSchema extends MixinDeclareSchema
             ->null()
             ;
 
-        $this->column('revision_updated_at')
-            ->datetime()
+        $this->column('revision_created_at')
+            ->timestamp()
             ->isa('DateTime')
             ->default(function() { 
                 return new \DateTime;
             });
 
-        $this->column('revision_created_at')
-            ->datetime()
+        $this->column('revision_updated_at')
+            ->timestamp()
             ->isa('DateTime')
-            ->default(function() { 
-                return new \DateTime;
-            });
+            ->default(new Raw('CURRENT_TIMESTAMP'));
+
 
         $this->belongsTo('root_revision', get_class($this->parentSchema), 'id', 'revision_root_id');
         $this->belongsTo('parent_revision', get_class($this->parentSchema), 'id', 'revision_parent_id');